pyproject.toml
[tool.poetry]
name = "google_ads_downloader"
version = "0.0.1"
description = "Free Google Ads Report Downloader"
authors = ["Geunseong Jung <dreamwayjgs@gmail.com>"]
repository = "https://github.com/dreamwayjgs/google-ads-downloader"
documentation = "https://dreamwayjgs.github.io/google-ads-downloader/"
readme = "README.md"
packages = [{ include = "google_ads_downloader", from = 'src' }]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
google-ads = "^27.0.0"
pandas = "^2.3.1"
inquirerpy = "^0.3.4"
openpyxl = "^3.1.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
deptry = "^0.16.2"
mypy = "^1.5.1"
pyright = "^1.1.382"
pre-commit = "^3.4.0"
tox = "^4.11.1"
jupyterlab = "^4.4.5"
tabulate = "^0.9.0"
pyinstaller = "^6.14.2"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.2.7"
mkdocstrings = { extras = ["python"], version = "^0.26.1" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
include = ["google_ads_downloader"]
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ['src']

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true
select = [
  # flake8-2020
  "YTT",
  # flake8-bandit
  "S",
  # flake8-bugbear
  "B",
  # flake8-builtins
  "A",
  # flake8-comprehensions
  "C4",
  # flake8-debugger
  "T10",
  # flake8-simplify
  "SIM",
  # isort
  "I",
  # mccabe
  "C90",
  # pycodestyle
  "E",
  "W",
  # pyflakes
  "F",
  # pygrep-hooks
  "PGH",
  # pyupgrade
  "UP",
  # ruff
  "RUF",
  # tryceratops
  "TRY",
]
ignore = [
  # LineTooLong
  "E501",
  # DoNotAssignLambda
  "E731",
]

[tool.ruff.format]


branch = true
source = ["google_ads_downloader"]
[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
